[css-grid] Fix grid container sizing under min-content height
The spec is quite clear
(https://drafts.csswg.org/css-sizing/#valdef-width-min-content):
  "min-content
     If specified for the inline axis, use the min-content inline size;
     otherwise behaves as the property’s initial value."

So if a grid container has "height: min-content" it should behave
the same than with "height: auto".

The patch removes ComputeIntrinsicLogicalContentHeightUsing() in
LayoutGrid as we don't need a custom one anymore.
We can also get rid of min_content_height_ and max_content_height_
attributes that were only used for this logic.

BUG=906039
TEST=fast/css-grid-layout/grid-track-sizing-with-orthogonal-flows.html
TEST=fast/css-grid-layout/maximize-tracks-definite-indefinite-height.html

Change-Id: Ia67d1dac16fbb04e3225115c3f34c67975dc6f94